// src/types/chat/api.ts
export interface Attachment {
  id: string;
  message_id: string;
  file_type: 'image' | 'video' | 'audio' | 'file' | 'location' | 'ig_reel';
  extension: string | null;
  data_url: string;
  thumb_url: string | null;
  file_size: number;
  fallback_title: string;
  coordinates_lat: number;
  coordinates_long: number;
  external_url?: string;
  transcribed_text?: string;
  meta?: Record<string, any>;
}